Moved: Benchmark only functions behind benchmark feature for split_565_color_endpoints
