+++ projects/dxt-lossless-transform-common/benches/split_color565_endpoints/sse2.rs
@@ -1,7 +1,7 @@
 #![allow(dead_code)]
 #![allow(unused_variables)]
 use criterion::{black_box, BenchmarkId};
-use dxt_lossless_transform_common::transforms::split_565_color_endpoints::bench::split_565_color_endpoints::{
+use dxt_lossless_transform_common::transforms::split_565_color_endpoints::bench::{
     sse2_shift_impl, sse2_shuf_impl, sse2_shuf_unroll2_impl,
 };
 use safe_allocator_api::RawAlloc;
