+++ projects/dxt-lossless-transform-common/src/transforms/split_565_color_endpoints/bench/ssse3.rs
